The modularization design of VizModel is insufficient. The user interface is bulky (a large number of public methods). The goal of extracting XML methods from this class was not only to solve the above smell, but also to ensure that this class adhered to the "single responsibility concept."
